fix: send UserType as numeric value (99) in JWT instead of string (Admin)

+++ MyIndustry.Identity.Domain/Service/AuthService.cs
@@ -155,7 +155,7 @@ public class AuthService : IAuthService
                 new Claim(JwtRegisteredClaimNames.Sub, user.UserName),
                 new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
                 new Claim(JwtRegisteredClaimNames.Email, user.Email),
-                new Claim("type", user.Type.ToString()),
+                new Claim("type", ((int)user.Type).ToString()),
                 new Claim("uid", user.Id)}
             .Union(userClaims)
             .Union(roleClaims);
